TimeInterpolatedGeoData.jl

TimeInterpolatedGeoData.CachedStack — Type

CachedStack(stack; loaded=Dict{Symbol,Any}())

A stack that wraps another Stack and stores any layers loaded from it in a Dict. This prevents loading them multiple times from disk.

It also means the are locked from the garbage collector if they are still in the Dict, which can cause problems with memory use. Use clear!(stack) to return to the unloaded state, freeing memory.

TimeInterpolatedGeoData.Cosine — Type

Cosine()

Similar to a Linear interpolator, but using sin to calculate the weights, as if it is cyclic between two values.

TimeInterpolatedGeoData.HypTan — Type

HypTan{S}()

Similar to a Linear interpolator, but using hyperbolic tangent to calculate the weights. It has a multiplier X that controls how long the curve is near each index. This is automatically scaled so the max and min values are always 0 and 1 for all X. At the limits, HypTan is essentially Linear and NoInterp.

HypTan{0} will error, instead uses Linear - which is identical.

TimeInterpolatedGeoData.InterpArray — Type

InterpArray(arrays::Tuple, fraction::Tuple)

InterplatedArray holds multiple arrays the will be interpolated when indexed. The fractions to include from each arrays are captured in the fractions argument. You may use as many arrays as required.

TimeInterpolatedGeoData.MinMaxInterpolator — Type

MinMaxIterpolator(times, interptype)

Specifies the time that correspond to the daily minimum and maximum.

Arguments

times: a length 2 NamedTuple where the keys correspond to the layer names for the minumum and maximum values, and the values specify the offset in hours at which they occur.

interptype: An Interpolations.jl InterpType. This package provides Cosine and HypTan.

TimeInterpolatedGeoData.interpseries — Method

interpseries(series::GeoSeries, dates; interptypes, step=step(dates))

Generate a an interpolated GeoSeries from series for each date/time in dates.

dates must define step if step kw is not passed in. An AbstractRange or a DimensionalData.jl Dimension can also be used.

Arguments

series: a GeoSeries of GeoStack, with a Regular Sampled Ti dimension.

dates: AbstractVector of dates

interptypes: a NamedTuple connecting a stack layer name with an Interpolations.jl InterpType.

TimeInterpolatedGeoData.interpstack — Method

interpstack(stacks, interptypes, frac)

Arguments

stack: a Tuple of AbstractGeoStack, usually of length 2.

interptypes: a NamedTuple connecting stack layer name with an Interpolations.jl InterpType. These are used for other layers used with the min/max layers.

frac: The fractional value, between 0 and 1, to interpolate to if the first stack is at 0 and the second stack is at 1.

Example

istack = interpstack((stack1, stack2), (a=BSpline(Linear()), b=BSpline(Cosine())), 0.25)

TimeInterpolatedGeoData.meanday_minmaxseries — Method

meanday_minmaxseries(series::GeoSeries, dates, interptypes, mm_interpolators::NamedTuple) => GeoSeries

Generate a GeoSeries of MinMixStack for one day in each period in series, with slices for every step in the day.

Arguments

series: a GeoSeries of GeoStack

dates: AbstractVector of dates

step: the step size of the intervals in the resulting GeoSeries.

interptypes: a NamedTuple connecting stack layer name with an Interpolations.jl InterpType. These are used for other layers used with the min/max layers.

mm_interpolators: a NamedTuple connecting new layer names (e.g. :temp) to a MinMaxFrac using other stack layers.

TimeInterpolatedGeoData.minmaxseries — Method

minmaxseries(series::GeoSeries, dates; interptypes, mm_interpolators::NamedTuple)

Generate a GeoSeries of MinMaxStack from series for each date/time in dates. Min/max interpolators are defined in the mm_interpolators, a NamedTuple of

dates must define step if step kw is not passed in. An AbstractRange or a DimensionalData.jl Dimension can also be used.

Arguments

series: a GeoSeries of GeoStack, with a Regular Sampled Ti dimension.

dates: AbstractVector of dates

interptypes: a NamedTuple connecting stack layer name with an Interpolations.jl InterpType. These are used for other layers used with the min/max layers.

mm_interpolators: a NamedTuple connecting new layer names (e.g. :temp) to a MinMaxFrac using other stack layers.

TimeInterpolatedGeoData.minmaxstack — Method

minmaxstack(stacks, interptypes, tfrac, mm_fracs)
minmaxstack(stacks, interptypes, tfrac, mm_interpolators, timestep)

An AbstractGeoStack that Provides interpolation for a value (often temperature) that usually has minimum and maximum values provided in separate layers. Other layers in the stack are interpolated, or not, as for interpstack.

Arguments

stack: a Tuple of AbstractGeoStack, usually of length 3.

interptypes: a NamedTuple connecting stack layer name with an Interpolations.jl InterpType. These are used for other layers used with the min/max layers.

tfrac: the fraction of the outer timestep this slice is taken at

mm_fracs: a NamedTuple connecting new layer names (e.g. :temp) to a MinMaxFrac using other stack layers.

mm_interpolators: a NamedTuple connecting new layer names (e.g. :temp) to a MinMaxFrac using other stack layers.

timestep: the size of the inner timestep, e.g Hour(1)